Rename some files and tests.
