+++ scripts/export-to-google-sheets.ts
@@ -4,8 +4,8 @@ import minimist from 'minimist';
 import { CollectedIncentive } from '../src/data/state_incentives';
 import { FILES, IncentiveFile } from './incentive-spreadsheet-registry';
 import { authorize } from './lib/auth-helper';
-import { generateDataModelTable } from './lib/google-data-model';
-import { collectedIncentiveToGoogleSheet } from './lib/google-sheets-exporter';
+import { generateDataModelTable } from './lib/google-sheets/data-model';
+import { collectedIncentiveToGoogleSheet } from './lib/google-sheets/incentives-sheet-exporter';
 import { FIELD_MAPPINGS, FIELD_METADATA } from './lib/spreadsheet-mappings';
 
 const INCENTIVE_SHEET_NAME = 'Incentives Data';
